- Added scroll bar to content div. - Changed session inactivity time out. - Resolved delete issue of application.

+++ Web/src/main/java/com/web/services/impl/ApplicationRCServiceImpl.java
@@ -1,10 +1,16 @@
 package com.web.services.impl;
 
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.data.services.CollegeProgramMapService;
+import com.data.entities.Application;
+import com.data.entities.Enums;
+import com.data.entities.StudentSelectedSubject;
 import com.data.services.ApplicationService;
+import com.data.services.CollegeProgramMapService;
+import com.data.services.StudentSelectedSubjectsDao;
 import com.web.services.ApplicationRCService;
 
 @Service("applicationRCService")
@@ -16,9 +22,22 @@ public class ApplicationRCServiceImpl implements ApplicationRCService {
 	@Autowired
 	ApplicationService applicationService;
 	
+	@Autowired
+	StudentSelectedSubjectsDao studentSelectedSubjectsDao;
+	
 	@Override
-	public void deleteAppliedCourse(Integer applicationId) {
-		applicationService.deleteById(applicationId);	
+	public Boolean deleteAppliedCourse(Integer applicationId) {
+		Application application = applicationService.getById(applicationId);
+		if(application != null) {
+			Enums.ApplicationStatus applicationStatus = Enums.ApplicationStatus.get(application.getApplicationStatus());
+			if(applicationStatus.getSequenceNo() <= Enums.ApplicationStatus.Partial.getSequenceNo()) { 
+				List<StudentSelectedSubject> studentSelectedSubjects = studentSelectedSubjectsDao.getSelectedSubjectOfApplication(applicationId);
+				studentSelectedSubjectsDao.deleteBySelectedSubjectList(studentSelectedSubjects);
+				applicationService.deleteById(applicationId);
+				return true;
+			}
+		}
+		return false;
 	}
 
 }
